Make create_dataset a reusable utility.
