+++ flocker/acceptance/integration/__init__.py
@@ -1,0 +1,5 @@
+# Copyright ClusterHQ Inc.  See LICENSE file for details.
+
+"""
+Tests that run a full Flocker cluster with external software.
+"""
